don't attempt to reopen a controller right after disconnect
i don't believe the previous behavior was desirable. imagine a scenario where the user has more than one controller connected. if the currently used controller gets erroneously disconnected, it'd immediately switch to the other controller and prevent reconnecting the first one until the second one is unplugged.

i think it's best to keep the controller used as consistent as possible throughout the plugin runtime. of course it would be best if you could simply choose the controller used, but i think this is the best solution for now